+++ ls/CHANGELOG.md
@@ -13,6 +13,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+[Unreleased]
+------------
+
+### Added
+
+-  Add support for `~~~~` prefix in script paths.
+   `~~~~` is an alias for the active script directory.
+
+
 [0.11.0] - 2026-05-18
 ---------------------